atualizei o botão em 2 componentes separados
