+++ src/components/atoms/Button/Button.tsx
@@ -1,31 +1,14 @@
-import classNames from "classnames";
-import React, {
-  type ComponentPropsWithRef,
-  type ForwardedRef,
-  forwardRef,
-} from "react";
+import type { ComponentPropsWithRef } from "react";
+import { forwardRef } from "react";
 
-export interface ButtonProps extends ComponentPropsWithRef<"button"> {
-  variant?: "outline" | "condensed";
-}
+export type ButtonProps = ComponentPropsWithRef<"button"> & {};
 
-export const Button = forwardRef(function Button(
-  { type = "button", variant, className, children, ...props }: ButtonProps,
-  ref: ForwardedRef<HTMLButtonElement>
-) {
-  return (
-    <button
-      type={type}
-      className={classNames(
-        "font-semibold p-3 rounded transition ease-in-out flex justify-center items-center",
-        variant === "outline" && "border border-primary-dark text-primary-dark",
-        variant === "condensed" && "bg-primary-dark text-white",
-        className
-      )}
-      {...props}
-      ref={ref}
-    >
-      {children}
-    </button>
-  );
-});
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
+  function Button({ children, type = "button", ...props }, ref) {
+    return (
+      <button type={type} {...props} ref={ref}>
+        {children}
+      </button>
+    );
+  }
+);
